requirements.txt
flask
flask-cors
numpy
opencv-python
pandas
pillow
